refactor: split data loaders into modules
